+++ Alloy/commands/compile/parsers/Alloy.Require.js
@@ -129,19 +129,31 @@ function parse(node, state, args) {
 		args.createArgs,
 		state
 	) + ')';
-	let parent = { symbol: args.symbol };
-	if (args.parent.symbol && !state.templateObject && !state.androidMenu && !state.insideContainer) {
+	// let parent = { symbol: args.symbol };
+	// if (args.parent.symbol && !state.templateObject && !state.androidMenu && !state.insideContainer) {
+	// 	code += ';\n' + args.symbol + '.setParent(' + args.parent.symbol + ');\n';
+	// 	parent = {
+	// 		symbol: args.symbol + '.getViewEx({recurse:true})'
+	// 	};
+	// } else if ( state.insideContainer ) {
+	// 	code += ';\n';
+	// 	parent = {
+	// 		symbol: args.symbol + '.getViewEx({recurse:true})'
+	// 	};
+	// } else {
+	// 	code += '.getViewEx({recurse:true});\n';
+	// }
+
+	let parent = {
+		symbol: args.symbol + '.getViewEx({recurse:true})'
+	};
+	if (args.parent.symbol && !state.templateObject && !state.androidMenu) {
 		code += ';\n' + args.symbol + '.setParent(' + args.parent.symbol + ');\n';
-		parent = {
-			symbol: args.symbol + '.getViewEx({recurse:true})'
-		};
-	} else if ( state.insideContainer ) {
-		code += ';\n';
-		parent = {
-			symbol: args.symbol + '.getViewEx({recurse:true})'
-		};
-	} else {
+	} else if (type === 'widget' && (node.parentNode && node.parentNode.nodeName === 'Alloy')) {
 		code += '.getViewEx({recurse:true});\n';
+		parent = { symbol: args.symbol };
+	} else {
+		code += ';\n';
 	}
 
 	return {